Extract OIDC to its own module
