+++ appnav/build.gradle.kts
@@ -42,6 +42,7 @@ dependencies {
     implementation(projects.libraries.architecture)
     implementation(projects.libraries.deeplink)
     implementation(projects.libraries.matrix.api)
+    implementation(projects.libraries.oidc.api)
     implementation(projects.libraries.preferences.api)
     implementation(projects.libraries.push.api)
     implementation(projects.libraries.pushproviders.api)
@@ -66,6 +67,7 @@ dependencies {
     testImplementation(libs.test.truth)
     testImplementation(libs.test.turbine)
     testImplementation(projects.libraries.matrix.test)
+    testImplementation(projects.libraries.oidc.impl)
     testImplementation(projects.libraries.push.test)
     testImplementation(projects.libraries.pushproviders.test)
     testImplementation(projects.features.networkmonitor.test)
