--- features/login/impl/src/main/kotlin/io/element/android/features/login/impl/oidc/customtab/DefaultOidcActionFlow.kt
@@ -1,44 +1,0 @@
-/*
- * Copyright (c) 2023 New Vector Ltd
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package io.element.android.features.login.impl.oidc.customtab
-
-import com.squareup.anvil.annotations.ContributesBinding
-import io.element.android.features.login.api.oidc.OidcAction
-import io.element.android.features.login.api.oidc.OidcActionFlow
-import io.element.android.libraries.di.AppScope
-import io.element.android.libraries.di.SingleIn
-import kotlinx.coroutines.flow.FlowCollector
-import kotlinx.coroutines.flow.MutableStateFlow
-import javax.inject.Inject
-
-@ContributesBinding(AppScope::class)
-@SingleIn(AppScope::class)
-class DefaultOidcActionFlow @Inject constructor() : OidcActionFlow {
-    private val mutableStateFlow = MutableStateFlow<OidcAction?>(null)
-
-    override fun post(oidcAction: OidcAction) {
-        mutableStateFlow.value = oidcAction
-    }
-
-    suspend fun collect(collector: FlowCollector<OidcAction?>) {
-        mutableStateFlow.collect(collector)
-    }
-
-    fun reset() {
-        mutableStateFlow.value = null
-    }
-}
