--- features/login/impl/src/main/kotlin/io/element/android/features/login/impl/oidc/webview/OidcNode.kt
@@ -1,54 +1,0 @@
-/*
- * Copyright (c) 2023 New Vector Ltd
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package io.element.android.features.login.impl.oidc.webview
-
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import com.bumble.appyx.core.modality.BuildContext
-import com.bumble.appyx.core.node.Node
-import com.bumble.appyx.core.plugin.Plugin
-import dagger.assisted.Assisted
-import dagger.assisted.AssistedInject
-import io.element.android.anvilannotations.ContributesNode
-import io.element.android.libraries.architecture.NodeInputs
-import io.element.android.libraries.architecture.inputs
-import io.element.android.libraries.di.AppScope
-import io.element.android.libraries.matrix.api.auth.OidcDetails
-
-@ContributesNode(AppScope::class)
-class OidcNode @AssistedInject constructor(
-    @Assisted buildContext: BuildContext,
-    @Assisted plugins: List<Plugin>,
-    presenterFactory: OidcPresenter.Factory,
-) : Node(buildContext, plugins = plugins) {
-    data class Inputs(
-        val oidcDetails: OidcDetails,
-    ) : NodeInputs
-
-    private val inputs: Inputs = inputs()
-    private val presenter = presenterFactory.create(inputs.oidcDetails)
-
-    @Composable
-    override fun View(modifier: Modifier) {
-        val state = presenter.present()
-        OidcView(
-            state = state,
-            modifier = modifier,
-            onNavigateBack = ::navigateUp,
-        )
-    }
-}
