--- features/login/impl/src/main/kotlin/io/element/android/features/login/impl/oidc/webview/OidcStateProvider.kt
@@ -1,39 +1,0 @@
-/*
- * Copyright (c) 2023 New Vector Ltd
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package io.element.android.features.login.impl.oidc.webview
-
-import androidx.compose.ui.tooling.preview.PreviewParameterProvider
-import io.element.android.libraries.architecture.AsyncAction
-import io.element.android.libraries.matrix.api.auth.OidcDetails
-
-open class OidcStateProvider : PreviewParameterProvider<OidcState> {
-    override val values: Sequence<OidcState>
-        get() = sequenceOf(
-            aOidcState(),
-            aOidcState().copy(requestState = AsyncAction.Loading),
-        )
-}
-
-fun aOidcState() = OidcState(
-    oidcDetails = aOidcDetails(),
-    requestState = AsyncAction.Uninitialized,
-    eventSink = {}
-)
-
-fun aOidcDetails() = OidcDetails(
-    url = "aUrl",
-)
